restructure class TMC_2209 for possible support of other drivers
